+++ src/localization/en/translation.json
@@ -74,6 +74,7 @@
     "songSendToTitle": "Send to...",
     "songRenameTitle": "Rename...",
     "songSearchInPlaylistTitle": "Search in Playlist",
+    "songStartRadio": "Start Radio",
     "songRemoveTitle": "Remove",
     "songRemoveNBanTitle": "Remove and BAN",
     "selectedSongs": "Selected songs"
