+++ src/localization/zhcn/translation.json
@@ -74,6 +74,7 @@
     "songSendToTitle": "送往...",
     "songRenameTitle": "改名...",
     "songSearchInPlaylistTitle": "在歌单内搜索",
+    "songStartRadio": "播放电台",
     "songRemoveTitle": "删除歌曲",
     "songRemoveNBanTitle": "删除并拉黑",
     "selectedSongs": "选择的歌"
